+++ js/modules/hist/TF1Painter.mjs
@@ -322,7 +322,7 @@ class TF1Painter extends TH1Painter {
       if (this._use_saved_points)
          return super.processTooltipEvent(pnt);
 
-      let ttrect = this.draw_g?.select('.tooltip_bin');
+      let ttrect = this.draw_g?.selectChild('.tooltip_bin');
 
       if (!this.draw_g || !pnt) {
          ttrect?.remove();