+++ js/changes.md
@@ -1,13 +1,13 @@
 # JSROOT changelog
 
-## Changed in dev
+## Changes in dev
 
 1. Correctly implement TH2 projections like MERCATOR or PARABOLIC
 2. Use https://github.com/georgealways/lil-gui/ instead of dat.GUI
 3. Let configure material and scene properties in geom control gui
 4. Upgrade three.js r151 -> r153
 5. Let toggle vertical/horizontal flag for color palette via context menu
-6. Provide "Bring to front" menu command for title/stats/palette/legend objects
+6. Provide "Bring to front" menu command for different objects like pave, box, marker, ...
 7. Handle "dark mode" in geom painter - automatically adjust background